Disable SSL config for MariaDB
In version 10.6.11, MariaDB started crashing if an incorrect SSL config is provided. Unfortunately, the default config on Ubuntu is incorrect, which means that MariaDB crashes immediately or cannot even install properly.

Puppet developers brought up this issue in https://github.com/puppetlabs/puppetlabs-mysql/issues/1509 and pointed to a fix that they've implemented. If `ssl => false` is passed in the option overrides for the mysql Puppet module, all other SSL config parameters are now excluded from `my.cnf`. 

I have confirmed that this change fixes the issue.